Refactoring base to package
This gives better overivew when working with a single service
