+++ objectcube/services/base/__init__.py
@@ -1,0 +1,6 @@
+from blob import BaseBlobService
+from concept import BaseConceptService
+from dimension import BaseDimensionService
+from object import BaseObjectService
+from plugin import BasePluginService
+from tag import BaseTagService
